requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
openai-whisper
torch
soundfile
numpy
ffmpeg-python
pydub